feat(capabilities): add space/content/decrypt capability
